AMPOPS-38 AMP 1.16.14-1.16.15 merged into AMP 1.17
git-svn-id: file:///svn/branches/AMP_1_17_RELEASE@17882 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/gis/action/GetFoundingDetails.java
@@ -506,7 +506,7 @@ public class GetFoundingDetails extends Action {
                                 get(key);
                         XML regionData = new XML("region");
                         regionData.addAttribute("reg-code", key);
-                        regionData.addAttribute("fundingCommitment",FormatHelper.formatNumber(ammount.getCommitment().intValue()));
+                        regionData.addAttribute("fundingCommitment",FormatHelper.formatNumber(ammount.getCommitment().doubleValue()));
                         regionData.addAttribute("fundingDisbursement",FormatHelper.formatNumber(ammount.getDisbursement().doubleValue()));
                         regionData.addAttribute("fundingExpenditure",FormatHelper.formatNumber(ammount.getExpenditure().doubleValue()));
                         root.addElement(regionData);
